+++ src/services/RecipeService.jsx
@@ -98,3 +98,80 @@
 //     return { success: false, error: error.message };
 //   }
 // };
+
+// RecipeService.jsx
+import { Config } from '../services/Config';
+
+const fetchRecipes = async () => {
+  try {
+    const response = await fetch(`${Config.API_URL}/recipes`);
+    if (!response.ok) {
+      throw new Error('Failed to fetch recipes');
+    }
+    return await response.json();
+  } catch (error) {
+    console.error('Error fetching recipes:', error.message);
+    throw error;
+  }
+};
+
+const addRecipe = async (newRecipe) => {
+  try {
+    const response = await fetch(`${Config.API_URL}/recipes`, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify(newRecipe),
+    });
+
+    if (!response.ok) {
+      throw new Error('Failed to add recipe');
+    }
+
+    return await response.json();
+  } catch (error) {
+    console.error('Error adding recipe:', error.message);
+    throw error;
+  }
+};
+
+const updateRecipe = async (id, updatedRecipe) => {
+  try {
+    const response = await fetch(`${Config.API_URL}/recipes/${id}`, {
+      method: 'PUT',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify(updatedRecipe),
+    });
+
+    if (!response.ok) {
+      throw new Error('Failed to update recipe');
+    }
+
+    return await response.json();
+  } catch (error) {
+    console.error('Error updating recipe:', error.message);
+    throw error;
+  }
+};
+
+const deleteRecipe = async (id) => {
+  try {
+    const response = await fetch(`${Config.API_URL}/recipes/${id}`, {
+      method: 'DELETE',
+    });
+
+    if (!response.ok) {
+      throw new Error('Failed to delete recipe');
+    }
+
+    return await response.json();
+  } catch (error) {
+    console.error('Error deleting recipe:', error.message);
+    throw error;
+  }
+};
+
+export { fetchRecipes, addRecipe, updateRecipe, deleteRecipe };
